pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-commit-vibe" 
version = "0.1.0"
description = "AI-powered conventional commit generator"
dependencies = [
    "google-genai",
]

[project.scripts]
ai-commit = "ai_commit:main"